+++ app/models/user.rb
@@ -106,6 +106,7 @@ class User < ApplicationRecord
     result[:menu_item_season_admin] = ph[:admin].present?
     result[:menu_item_analytics_admin] = ph[:admin].present?
     result[:menu_item_email_log_admin] = ph[:admin].present?
+    result[:menu_item_email_template_admin] = ph[:admin].present?
 
     # show permissions
     result[:show_league_index_admin] = ph[:admin].present? || ph[:sbk].present?