+++ config/routes.rb
@@ -224,6 +224,8 @@ Rails.application.routes.draw do
           get :incorrect_assignments, on: :collection
           get :next_lizenznummer, on: :collection
         end
+        resources :email_templates, only: %i[index]
+        patch 'email_templates', to: 'email_templates#update'
         resources :referee_qualification_types, only: %i[index create update destroy]
         resources :referee_license_levels, only: %i[index create update destroy]
         resources :penalty_codes, only: %i[index create update destroy]